style(phantun) refactor out common functions and constants
